backend/requirements.txt
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.18.0
firebase-admin==6.7.0
Flask==3.1.0
flask-cors==5.0.1
fsspec==2025.3.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1rc1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
setuptools==78.1.0
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tqdm==4.67.1
transformers==4.50.1
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
